backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
requests
python-dotenv
motor
simple-salesforce
apscheduler
pymongo
kaggle
sendgrid